backend/requirements.txt
# Django
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<5.0
django-filter>=24.3,<25.0
django-environ>=0.11,<1.0
django-celery-beat>=2.7,<3.0

# Auth
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9,<3.0

# Geo / PostGIS
django-contrib-gis
djangorestframework-gis>=1.1,<2.0

# Async tasks
celery>=5.4,<6.0
redis>=5.1,<6.0

# Server
gunicorn>=22.0,<23.0

# Utilities
python-decouple>=3.8,<4.0
requests>=2.32,<3.0

# Rate limiting
django-ratelimit>=4.1,<5.0

# Logging
structlog>=24.4,<25.0

# API documentation
drf-spectacular>=0.28,<1.0

# Testing
pytest-django>=4.9,<5.0
factory-boy>=3.3,<4.0
